feat: 콘페티 로직 분리
